LayoutTests/imported/w3c/web-platform-tests/mixed-content/gen/sharedworker-classic.http-rp/unset/websocket.https-expected.txt


FAIL Mixed-Content: Expects allowed for websocket to same-wss origin and no-redirect redirection from https context. assert_equals: The resource request should be 'allowed'. expected "allowed" but got "blocked"
PASS Mixed-Content: Expects blocked for websocket to cross-ws origin and no-redirect redirection from https context.
PASS Mixed-Content: Expects blocked for websocket to same-ws origin and no-redirect redirection from https context.
